Fix bug in calculation of pretty-printer output width for linebreak
linebreak is line("") which means that if the line break is omitted it is replaced
by nothing. There was an error in the line method which gave all line breaks
a width of 1 if they were omitted, even these ones. We now use the actual
width of the replacement. A flow-on effect was that a couple of tests and the
fillcat combinator had to be fixed since they had been working around this
bug and didn't work when it was fixed.