Add support for singletons in node duplication, don't duplicate them (issue 70)
This fix addresses a problem whereby a deepclone of a list would crash when it got
to the Nil since the class that represents that singleton does not have a constructor.
The correct solution is to not clone singletons. In Scala that includes values that
are declared as objects and case objects.

The new version returns the same instance if a singleton is to be duplicated, only
falling back on the constructor if the instance is not a singleton.

Ordinarily this bug does not occur when using the generic traversals (and hence
the dup method) because the traversals short-circuit situations where there are
no children of a node. Nevertheless, the fix also applies to dup in case there is
ever a situation where a node is to be duplicated directly (as is done for leaves
during a deepclone).